+++ packages/docs/.vitepress/theme/index.js
@@ -1,4 +1,5 @@
 import DefaultTheme from 'vitepress/theme';
+import TwoslashFloatingVue from '@shikijs/vitepress-twoslash/client';
 import PreviewIframe from './components/PreviewIframe.vue';
 import './custom.scss';
 
@@ -7,5 +8,6 @@ export default {
   extends: DefaultTheme,
   enhanceApp({ app }) {
     app.component('PreviewIframe', PreviewIframe);
+    app.use(TwoslashFloatingVue);
   },
 };