+++ prettier.config.js
@@ -1,1 +1,14 @@
-export { default } from '@studiometa/prettier-config';
+import config from '@studiometa/prettier-config';
+
+export default {
+  ...config,
+  overrides: [
+    ...config.overrides,
+    {
+      files: 'packages/docs/**/*.md',
+      options: {
+        printWidth: 78,
+      },
+    },
+  ],
+};