+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file. The format 
 
 ## [Unreleased]
 
+## [v3.0.3](https://github.com/studiometa/js-toolkit/compare/3.0.2..3.0.3) (2025-05-09)
+
+### Changed
+
+- **useResize:**
+  - improve types ([4ee26087](https://github.com/studiometa/js-toolkit/commit/4ee26087))
+  - improve breakpoint read performance ([#589](https://github.com/studiometa/js-toolkit/pull/589), [4da5ed4c](https://github.com/studiometa/js-toolkit/commit/4da5ed4c))
+- **useDrag:** refactor MODES to the service props ([#618](https://github.com/studiometa/js-toolkit/pull/618), [88c8bb54](https://github.com/studiometa/js-toolkit/commit/88c8bb54))
+- Reuse the existing `domScheduler` ([#607](https://github.com/studiometa/js-toolkit/issues/607), [#608](https://github.com/studiometa/js-toolkit/pull/608), [ebf96d2b](https://github.com/studiometa/js-toolkit/commit/ebf96d2b))
+
 ## [v3.0.2](https://github.com/studiometa/js-toolkit/compare/3.0.1..3.0.2) (2025-04-09)
 
 ### Changed